fixed this compile error on htmltotext.d:
htmltotext.d(64): Error: function arsd.htmltotext.repeat (string s, int num) is not callable using argument types (string, ulong)
htmltotext.d(69): Error: function arsd.htmltotext.repeat (string s, int num) is not callable using argument types (string, ulong)

+++ htmltotext.d
@@ -4,7 +4,7 @@ import arsd.dom;
 import std.string;
 import std.array : replace;
 
-string repeat(string s, int num) {
+string repeat(string s, ulong num) {
 	string ret;
 	foreach(i; 0 .. num)
 		ret ~= s;
